requirements.txt
basicsr==1.4.2
gfpgan==1.3.8
numpy<2.0.0
Pillow==11.0.0
PyYAML==6.0.2
realesrgan==0.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.6
